Upgrade typescript and turn on noImplicitAny (#49)
# Background

This is PR 1 of 2 for rhttps://github.com/dbt-labs/dbt-cloud/issues/4032

We would like to adopt `noImplicitAny` to enable the TS compiler to provide us more guarantees, ultimately resulting in fewer bugs making it to production.

Updating fui is not _necessary_ for us to adopt `noImplicitAny` in dbt cloud. But this will 
1. Save us from having to define custom `.d.ts` files in dbt-cloud to compensate for ts errors in fui
2. Provide type safety benefits to fui as well 🎉 

# Work done
- Upgraded TypeScript to latest version
- Fixed type errors
- Formatted all code with Prettier so that `npm run lint` passes without error
- Upgraded storybook
  - Storybook commands were failing due to TS errors in a nested dependency. Upgrading storybook resolved this.
  - However, with storybook upgraded, storybook commands failed due to React being missing. I specified React as a peer dependency, and also installed it as a dev dependency for storybook. 

There are more details on this in some of the details of the commit messages along the way. 

# Acceptance testing
- [x] Visually and behaviorally diff the components in production storybook versus this PR's storybook.